// apps/api-lavperform/src/whatsapp/uazapi/application/dto/create-instance.dto.ts
export class CreateInstanceDto {
  name: string;
  systemName?: string;
  adminField01?: string;
  adminField02?: string;
  fingerprintProfile?: string;
  browser?: string;
}

export class CreateInstanceResponseDto {
  response: string;
  instance: {
    id: string;
    token: string;
    status: string;
    paircode: string;
    qrcode: string;
    name: string;
    profileName: string;
    profilePicUrl: string;
    isBusiness: boolean;
    plataform: string;
    systemName: string;
    owner: string;
    lastDisconnect: string;
    lastDisconnectReason: string;
    adminField01: string;
    openai_apikey: string;
    chatbot_enabled: boolean;
    chatbot_ignoreGroups: boolean;
    chatbot_stopConversation: string;
    chatbot_stopMinutes: number;
    created: string;
    updated: string;
    currentPresence: string;
  };
  connected: boolean;
  loggedIn: boolean;
  name: string;
  token: string;
  info: string;
}