// apps/api-lavperform/src/whatsapp/uazapi/application/dto/connect-instance.dto.ts
export class ConnectInstanceResponseDto {
  connected: boolean;
  loggedIn: boolean;
  jid: string;
  instance: {
    id: string;
    token: string;
    status: string;
    paircode: string;
    qrcode: string;
    name: string;
    profileName: string;
    profilePicUrl: string;
    isBusiness: boolean;
    plataform: string;
    systemName: string;
    owner: string;
    adminField01: string; 
    openai_apikey: string;
    chatbot_enabled: boolean;
    chatbot_ignoreGroups: boolean;
    chatbot_stopConversation: string;
    chatbot_stopMinutes: number;
    created: string;
    updated: string;
    currentPresence: string;
  }
}